Restore Timer::Duration to return float

+++ onnxruntime/contrib_ops/rocm/bert/util.cc
@@ -27,11 +27,11 @@ void Timer::End() {
   HIP_CHECK(hipEventSynchronize(end_));
 }
 
-double Timer::Duration() {
+float Timer::Duration() {
   float time;
   // time is in ms with a resolution of 1 us
   HIP_CHECK(hipEventElapsedTime(&time, start_, end_));
-  return static_cast<double>(time);
+  return time;
 }
 
 Timer::~Timer() {
